chore(Domain): 유효성 검사 파일 위치 변경
- 통합모듈 -> 업무별 도메인
