+++ tablegroup/src/test/java/kitchenpos/tablegroup/domain/TableGroupValidatorTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.tablegroup.domain;
 
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyLong;
@@ -15,14 +15,12 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import kitchenpos.common.vo.TableGroupId;
-import kitchenpos.order.application.OrderService;
 import kitchenpos.table.application.TableService;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTables;
-import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.table.dto.OrderTableDto;
 import kitchenpos.tablegroup.dto.TableGroupDto;
-import kitchenpos.validation.TableGroupValidatorImpl;
+import kitchenpos.order.application.OrderService;
 import kitchenpos.order.exception.HasNotCompletionOrderException;
 import kitchenpos.table.exception.HasOtherTableGroupException;
 import kitchenpos.table.exception.NotEmptyOrderTableException;
@@ -38,7 +36,7 @@ public class TableGroupValidatorTest {
     TableService tableService;
 
     @InjectMocks
-    TableGroupValidatorImpl tableGroupValidator;
+    TableGroupValidator tableGroupValidator;
 
     @DisplayName("그룹해제시 주문상태가 전부 계산완료가아니면 예외가 발생된다.")
     @Test
